requirements.txt
avro-python3
requests
confluent-kafka
